add throw exception

+++ src/modules/PoC/services/poc.service.ts
@@ -1,27 +1,58 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { PoCEntity } from '../entities/poc.entity';
 import { PoCRepository } from '../repositories/poc.repository';
 
 @Injectable()
 export class PoCService {
   constructor(private poCRepository: PoCRepository) {}
-  getData(): Promise<PoCEntity[]> {
+  async getData(): Promise<PoCEntity[]> {
     return this.poCRepository.find({ select: ['id', 'username', 'status'] });
   }
 
-  createData(dataEntity: PoCEntity): Promise<PoCEntity> {
+  async createData(dataEntity: PoCEntity): Promise<PoCEntity> {
+    const userExistedQuery = await this.poCRepository.findOne({
+      where: { username: dataEntity.username },
+    });
+    if (!userExistedQuery) {
+      throw new BadRequestException();
+    }
     return this.poCRepository.save(dataEntity);
   }
 
-  getDataById(PoCId: number): Promise<PoCEntity[]> {
-    return this.poCRepository.find({ where: { id: PoCId } });
+  async getDataById(PoCId: number): Promise<PoCEntity> {
+    const userExistQuery = await this.poCRepository.findOne({
+      where: { id: PoCId },
+    });
+    if (!userExistQuery) {
+      throw new BadRequestException();
+    }
+    return userExistQuery;
   }
 
-  updateDataById(PoCId: number, dataEntity: PoCEntity) {
-    return this.poCRepository.update({ id: PoCId }, dataEntity);
+  async updateDataById(
+    PoCId: number,
+    dataEntity: PoCEntity
+  ): Promise<PoCEntity> {
+    const userExistedQuery = await this.poCRepository.findOne({
+      where: { id: PoCId },
+    });
+    const userExistedQueryByUsername = await this.poCRepository.findOne({
+      where: { username: dataEntity.username },
+    });
+    if (!userExistedQuery || !userExistedQueryByUsername) {
+      throw new BadRequestException();
+    }
+    this.poCRepository.update({ id: PoCId }, dataEntity);
+    return this.poCRepository.findOne({ where: { id: PoCId } });
   }
 
-  removeDataById(PoCId: number) {
+  async removeDataById(PoCId: number) {
+    const userExistedQuery = await this.poCRepository.findOne({
+      where: { id: PoCId },
+    });
+    if (!userExistedQuery) {
+      throw new BadRequestException();
+    }
     return this.poCRepository.delete(PoCId);
   }
 }
